Add string id type validation
This will allow us more flexibility when validating a string
representing an ID. Previously we were just checking that the ID was in
fact a string. This change introduces a new validation type that fits a
general case for all Keystone resources. This functionality defaults to
a regular expression but it can be strengthened through the validation
configuration group available in the keystone.conf.sample.

Change-Id: I4878a1c718159304960fc7c40c1a80b8599c5657
bp: api-validation