Adds backend check to setup of LDAP tests
I was finding that by making tweaks in the setup ordering that I was
getting the wrong backends for all (or parts) of a test. This extra
check ensures that I am not making things worse while refactoring.

The way we split setup among inheritance hierarchies makes knowing what
you're getting incredibly difficult.

Change-Id: I025c9a9e89a57bcff0a72cc1e584c3114671b3b3